# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Interop Backend Service API Documentation",
        "title": "Interop Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/blocks": {
            "get": {
                "description": "Get a paginated list of sync blocks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blocks"
                ],
                "summary": "Get sync blocks with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default: 10, max: 100)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": 200, \"msg\": \"success\", \"data\": {\"list\": []schema.SyncBlock, \"total\": int64, \"page\": int, \"pageSize\": int}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/blocks/relayedMessage/{hash}": {
            "get": {
                "description": "Get relayed message details using message hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Get relayed message by hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message Hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": 200, \"msg\": \"success\", \"data\": schema.SyncEvent",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/blocks/sentMessage/{hash}": {
            "get": {
                "description": "Get sent message details using message hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Get sent message by hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message Hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": 200, \"msg\": \"success\", \"data\": schema.SyncEvent",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/blocks/{blockNumber}/executingMessage": {
            "get": {
                "description": "Get all executing messages for a specific block number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Get executing messages by block number",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Block Number",
                        "name": "blockNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\": 200, \"msg\": \"success\", \"data\": []schema.SyncEvent",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "code\": 400, \"msg\": \"Invalid block number\", \"data\": nil",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    }
}